// packages/nestjskit__s3/src/s3.interface.ts
import type { ModuleMetadata, Type } from "@nestjs/common";

import type { S3ClientConfig } from "@aws-sdk/client-s3";

/** Known AWS regions offer completion; custom S3-compatible regions remain valid. */
export type S3Region =
  | "us-east-1"
  | "us-east-2"
  | "us-west-1"
  | "us-west-2"
  | "af-south-1"
  | "ap-east-1"
  | "ap-east-2"
  | "ap-south-1"
  | "ap-south-2"
  | "ap-northeast-1"
  | "ap-northeast-2"
  | "ap-northeast-3"
  | "ap-southeast-1"
  | "ap-southeast-2"
  | "ap-southeast-3"
  | "ap-southeast-4"
  | "ap-southeast-5"
  | "ap-southeast-6"
  | "ap-southeast-7"
  | "ca-central-1"
  | "ca-west-1"
  | "eu-central-1"
  | "eu-central-2"
  | "eu-west-1"
  | "eu-west-2"
  | "eu-west-3"
  | "eu-south-1"
  | "eu-south-2"
  | "eu-north-1"
  | "il-central-1"
  | "mx-central-1"
  | "me-south-1"
  | "me-central-1"
  | "sa-east-1"
  | (string & {});

export type ModuleOptions = Omit<S3ClientConfig, "region"> & {
  region?: S3Region | Exclude<S3ClientConfig["region"], string | undefined>;
};

export interface ExtraModuleOptions {
  /**
   * alias for module
   * @default ''
   */
  alias?: string;

  /**
   * global module
   * @default false
   */
  global?: boolean;
}

export interface ModuleOptionsFactory {
  create(): Promise<ModuleOptions> | ModuleOptions;
}

export interface AsyncModuleOptions extends Pick<ModuleMetadata, "imports"> {
  inject?: any[];
  useClass?: Type<ModuleOptionsFactory>;
  useExisting?: Type<ModuleOptionsFactory>;
  useFactory?: (...args: any[]) => Promise<ModuleOptions> | ModuleOptions;
}